api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "uvicorn[standard]>=0.35.0",
    "ably>=2.0.0",
    "sqlmodel>=0.0.24",
    "psycopg2-binary>=2.9.9",
    "alembic>=1.13.0",
    "python-dotenv>=1.1.1",
]
